Reflection-based api test cases are  moved to Cases/Reflectons directory
