+++ src/insights/FastEnum.UnitTests/Cases/Reflections/BasicTests.cs
@@ -12,7 +12,7 @@ using FastEnumUtility.UnitTests.Models;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace FastEnumUtility.UnitTests.Cases;
+namespace FastEnumUtility.UnitTests.Cases.Reflections;
