+++ src/insights/FastEnum.UnitTests/Cases/Reflections/BasicTests.tt
@@ -30,7 +30,7 @@ using FastEnumUtility.UnitTests.Models;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace FastEnumUtility.UnitTests.Cases;
+namespace FastEnumUtility.UnitTests.Cases.Reflections;
